+++ Frends.AmazonS3.DownloadObject/CHANGELOG.md
@@ -1,8 +1,11 @@
 # Changelog
 
-## [1.0.1] - 2022-12-02
+## [2.0.0] - 2022-12-02
 ### Modified
-- Fixing memory leaks. 
+- Memory leak fix.
+- SingleResultObject reform:
+	- New parameters: Overwritten, SourceDeleted, Info.
+	- Removed parameter: ObjectData (replaced by Info). 
 
 ## [1.0.0] - 2022-05-09
 ### Added